add dlt_tests package and move one fixture for testing
