+++ packages/dlt_tests/README.md
@@ -1,0 +1,30 @@
+<h1 align="center">
+    <strong>data load tool (dlt) pytest plugin & test helpers</strong>
+</h1>
+<p align="center">
+pytest plugin and test helpers for users and developers of dlt (https://dlthub.com/docs/intro). This package is currently in early development and considered an alpha.
+</p>
+
+
+<h3 align="center">
+
+🚀 Join our thriving community of likeminded developers and build the future together!
+
+</h3>
+
+<div align="center">
+  <a target="_blank" href="https://dlthub.com/community" style="background:none">
+    <img src="https://img.shields.io/badge/slack-join-dlt.svg?labelColor=191937&color=6F6FF7&logo=slack" style="width: 260px;"  />
+  </a>
+</div>
+<div align="center">
+  <a target="_blank" href="https://pypi.org/project/dlt_tests/" style="background:none">
+    <img src="https://img.shields.io/pypi/v/dlt_tests?labelColor=191937&color=6F6FF7">
+  </a>
+  <a target="_blank" href="https://pypi.org/project/dlt_tests/" style="background:none">
+    <img src="https://img.shields.io/pypi/pyversions/dlt_tests?labelColor=191937&color=6F6FF7">
+  </a>
+  <a target="_blank" href="https://pypi.org/project/dlt_tests/" style="background:none">
+    <img src="https://img.shields.io/pypi/dm/dlt_tests?labelColor=191937&color=6F6FF7">
+  </a>
+</div>
