+++ packages/dlt_tests/dlt_tests/destinations/__init__.py
@@ -1,0 +1,1 @@
+# test destination setup DestinationTestConfiguration etc.
