+++ pyproject.toml
@@ -77,6 +77,7 @@ workspace = ["dlt[workspace]"]
 # we use those deps in common tests to make sure they work without extra dependencies so if you add them
 # here you'll break everything
 dev = [
+    "dlt_tests",
     "requests-mock>=1.10.0,<2",
     "types-click>=7.1.8,<8",
     "sqlfluff>=2.3.2,<3",
@@ -196,6 +197,7 @@ docs = [
 
 [tool.uv.sources]
 dlt = { workspace = true }
+dlt_tests = { workspace = true }
 flake8-encodings = { git = "https://github.com/dlt-hub/flake8-encodings.git", branch = "disable_jedi_support" }
 
 [tool.uv.workspace]
