+++ tests/load/pipeline/conftest.py
@@ -1,1 +1,1 @@
-from tests.pipeline.utils import drop_dataset_from_env
+from dlt_tests.fixtures import drop_dataset_from_env
