+++ tests/pipeline/conftest.py
@@ -6,4 +6,4 @@ from tests.utils import (
     test_storage,
 )
 from tests.common.configuration.utils import environment, toml_providers
-from tests.pipeline.utils import drop_dataset_from_env
+from dlt_tests.fixtures import drop_dataset_from_env
